Reworking power expressions
